MMCore: fix compiler warnings
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16581 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ MMCore/CircularBuffer.cpp
@@ -366,7 +366,7 @@ const mm::ImgBuffer* CircularBuffer::GetNthFromTopImageBuffer(long n,
 
    long targetIndex = insertIndex_ - n - 1L;
    while (targetIndex < 0)
-      targetIndex += frameArray_.size();
+      targetIndex += (long) frameArray_.size();
    targetIndex %= frameArray_.size();
 
    return frameArray_[targetIndex].FindImage(channel);
